requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pymongo==4.10.1
pytz==2024.2
six==1.16.0
Werkzeug==3.0.4
python-dotenv
googlemaps
sentence_transformers
faiss-cpu
numpy<2
gunicorn
memory-profiler
# Added for web scraping and agent functionality
beautifulsoup4>=4.9.3 # For parsing HTML
requests>=2.25.1 # For making HTTP requests
playwright>=1.20 # For JS-heavy sites (optional, if bs4 is not enough)
openai>=1.0.0 # For OpenAI API (product categorization)
langchain>=0.1.0 # If using Langchain specific features (alternative for OpenAI client)
boto3>=1.20.0 # AWS SDK for Python (for S3)
